Move interface on a new api contracts package
