+++ src/Monofony/Contracts/Api/phpstan.neon
@@ -1,0 +1,6 @@
+parameters:
+    paths:
+        - ./
+    excludes_analyse:
+        - 'vendor/**.php'
+        - 'spec/**.php'
